feat: 마이페이지 상단부 UI 구현
